cleanup

+++ src/scripts/devspace/clean.sh
@@ -1,5 +1,9 @@
+#!/usr/bin/env zsh
 containerid=$(docker ps -q -f name=kindest-argo-cd_devcontainer-kindest)
-docker rm -f $containerid -f
+if [ ! (-z "$containerid") ]
+then
+    docker rm -f $containerid -f
+fi
 docker container prune -f
 docker image prune -a -f
 docker network prune -f
